+++ python/pynessie/model.py
@@ -16,6 +16,10 @@ class Contents:
 
     id: str = desert.ib(fields.Str())
 
+    def requires_expected_state(self: "Contents") -> bool:
+        """Checks whether this Contents object requires the "expected" state to be provided for Put operations."""
+        return False
+
     def pretty_print(self: "Contents") -> str:
         """Print out for cli."""
         pass
@@ -26,11 +30,15 @@ class IcebergTable(Contents):
     """Dataclass for Nessie Contents."""
 
     metadata_location: str = desert.ib(fields.Str(data_key="metadataLocation"))
-    snapshot_id: int = desert.ib(fields.Int(data_key="snapshotId"))
+    id_generators: str = desert.ib(fields.Str(data_key="idGenerators"))
+
+    def requires_expected_state(self: "IcebergTable") -> bool:
+        """Returns True - expected state should be provided for Put operations on Iceberg tables."""
+        return True
 
     def pretty_print(self: "IcebergTable") -> str:
         """Print out for cli."""
-        return "Iceberg table:\n\tmetadata-location:{}\n\tsnapshot-id:{}".format(self.metadata_location, self.snapshot_id)
+        return "Iceberg table:\n\tmetadata-location:{}\n\tid-generators:{}".format(self.metadata_location, self.id_generators)
 
 
 IcebergTableSchema = desert.schema_class(IcebergTable)
@@ -114,6 +122,7 @@ class Put(Operation):
     """Single Commit Operation."""
 
     contents: Contents = desert.ib(fields.Nested(ContentsSchema))
+    expectedContents: Optional[Contents] = attr.ib(default=None, metadata=desert.metadata(fields.Nested(ContentsSchema)))
 
 
 PutOperationSchema = desert.schema_class(Put)