+++ servers/jax-rs-tests/src/main/java/org/projectnessie/jaxrs/AbstractResteasyTest.java
@@ -33,7 +33,6 @@ import org.projectnessie.model.Contents;
 import org.projectnessie.model.ContentsKey;
 import org.projectnessie.model.IcebergTable;
 import org.projectnessie.model.ImmutableBranch;
-import org.projectnessie.model.ImmutableIcebergTable;
 import org.projectnessie.model.ImmutableOperations;
 import org.projectnessie.model.ImmutablePut;
 import org.projectnessie.model.LogResponse;
@@ -77,7 +76,7 @@ public class AbstractResteasyTest {
         newReference,
         rest().get("trees/tree/test").then().statusCode(200).extract().as(Branch.class));
 
-    IcebergTable table = IcebergTable.of("/the/directory/over/there", -1L);
+    IcebergTable table = IcebergTable.of("/the/directory/over/there", "x");
 
     Branch commitResponse =
         rest()
@@ -103,21 +102,13 @@ public class AbstractResteasyTest {
       updates[i] =
           ImmutablePut.builder()
               .key(ContentsKey.of("item", Integer.toString(i)))
-              .contents(
-                  ImmutableIcebergTable.builder()
-                      .metadataLocation("/the/directory/over/there/" + i)
-                      .snapshotId(-1L)
-                      .build())
+              .contents(IcebergTable.of("/the/directory/over/there/" + i, "x"))
               .build();
     }
     updates[10] =
         ImmutablePut.builder()
             .key(ContentsKey.of("xxx", "test"))
-            .contents(
-                ImmutableIcebergTable.builder()
-                    .metadataLocation("/the/directory/over/there/has/been/moved")
-                    .snapshotId(-1L)
-                    .build())
+            .contents(IcebergTable.of("/the/directory/over/there/has/been/moved", "x"))
             .build();
 
     Reference branch = rest().get("trees/tree/test").as(Reference.class);
@@ -143,12 +134,7 @@ public class AbstractResteasyTest {
     Assertions.assertEquals(updates[10].getContents(), res.body().as(Contents.class));
 
     IcebergTable currentTable = table;
-    table =
-        ImmutableIcebergTable.builder()
-            .from(table)
-            .metadataLocation("/the/directory/over/there/has/been/moved/again")
-            .snapshotId(-1L)
-            .build();
+    table = IcebergTable.of("/the/directory/over/there/has/been/moved/again", "x", table.getId());
 
     Branch b2 = rest().get("trees/tree/test").as(Branch.class);
     rest()
@@ -251,10 +237,10 @@ public class AbstractResteasyTest {
                 (expectedMetadataUrl != null)
                     ? Put.of(
                         ContentsKey.of(contentsKey),
-                        IcebergTable.of(metadataUrl, -1L, contentsId),
-                        IcebergTable.of(expectedMetadataUrl, -1L, contentsId))
+                        IcebergTable.of(metadataUrl, "x", contentsId),
+                        IcebergTable.of(expectedMetadataUrl, "x", contentsId))
                     : Put.of(
-                        ContentsKey.of(contentsKey), IcebergTable.of(metadataUrl, -1L, contentsId)))
+                        ContentsKey.of(contentsKey), IcebergTable.of(metadataUrl, "x", contentsId)))
             .commitMeta(CommitMeta.builder().author(author).message("").build())
             .build();
     return rest()