+++ servers/quarkus-server/src/main/java/org/projectnessie/server/providers/RocksVersionStoreFactory.java
@@ -19,11 +19,11 @@ import static org.projectnessie.server.config.VersionStoreConfig.VersionStoreTyp
 
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
+import org.projectnessie.server.config.QuarkusVersionStoreAdvancedConfig;
 import org.projectnessie.services.config.ServerConfig;
 import org.projectnessie.versioned.StoreWorker;
 import org.projectnessie.versioned.VersionStore;
 import org.projectnessie.versioned.persist.adapter.DatabaseAdapter;
-import org.projectnessie.versioned.persist.nontx.NonTransactionalDatabaseAdapterConfig;
 import org.projectnessie.versioned.persist.rocks.RocksDatabaseAdapterFactory;
 import org.projectnessie.versioned.persist.rocks.RocksDbInstance;
 import org.projectnessie.versioned.persist.store.PersistVersionStore;
@@ -32,9 +32,8 @@ import org.projectnessie.versioned.persist.store.PersistVersionStore;
 @StoreType(ROCKS)
 @Dependent
 public class RocksVersionStoreFactory implements VersionStoreFactory {
-
   @Inject RocksDbInstance rocksDbInstance;
-  @Inject NonTransactionalDatabaseAdapterConfig config;
+  @Inject QuarkusVersionStoreAdvancedConfig config;
 
   @Override
   public <VALUE, METADATA, VALUE_TYPE extends Enum<VALUE_TYPE>>