+++ servers/quarkus-server/src/test/java/org/projectnessie/server/error/ITNativeNessieError.java
@@ -50,7 +50,7 @@ public class ITNativeNessieError {
   @Test
   void testNullParamViolation() {
     ContentsKey k = ContentsKey.of("a");
-    IcebergTable t = IcebergTable.of("path1", -1L);
+    IcebergTable t = IcebergTable.of("path1", "x");
     assertEquals(
         "Bad Request (HTTP/400): commitMultipleOperations.hash: must not be null",
         assertThrows(