+++ tools/apprunner-maven-plugin/src/main/java/org/projectnessie/quarkus/maven/QuarkusApp.java
@@ -21,9 +21,8 @@ import io.quarkus.bootstrap.app.QuarkusBootstrap;
 import io.quarkus.bootstrap.app.QuarkusBootstrap.Mode;
 import io.quarkus.bootstrap.app.RunningQuarkusApplication;
 import io.quarkus.bootstrap.app.StartupAction;
-import io.quarkus.bootstrap.model.AppArtifact;
 import io.quarkus.bootstrap.model.AppArtifactCoords;
-import io.quarkus.bootstrap.model.AppModel;
+import io.quarkus.bootstrap.model.ApplicationModel;
 import io.quarkus.bootstrap.resolver.BootstrapAppModelResolver;
 import io.quarkus.bootstrap.resolver.maven.MavenArtifactResolver;
 import java.lang.reflect.Method;
@@ -70,16 +69,9 @@ public class QuarkusApp implements AutoCloseable {
       String appArtifactId,
       Properties applicationProperties)
       throws MojoExecutionException {
-    final AppArtifactCoords appCoords = AppArtifactCoords.fromString(appArtifactId);
-    final AppArtifact appArtifact =
-        new AppArtifact(
-            appCoords.getGroupId(),
-            appCoords.getArtifactId(),
-            appCoords.getClassifier(),
-            appCoords.getType(),
-            appCoords.getVersion());
+    AppArtifactCoords appCoords = AppArtifactCoords.fromString(appArtifactId);
 
-    final AppModel appModel;
+    ApplicationModel appModel;
     try {
       MavenArtifactResolver resolver =
           MavenArtifactResolver.builder()
@@ -93,10 +85,10 @@ public class QuarkusApp implements AutoCloseable {
           new BootstrapAppModelResolver(resolver)
               .setDevMode(false)
               .setTest(false)
-              .resolveModel(appArtifact);
+              .resolveModel(appCoords);
     } catch (Exception e) {
       throw new MojoExecutionException(
-          "Failed to resolve application model " + appArtifact + " dependencies", e);
+          "Failed to resolve application model " + appCoords + " dependencies", e);
     }
 
     return newApplication(
@@ -120,7 +112,10 @@ public class QuarkusApp implements AutoCloseable {
    * @throws MojoExecutionException if an error occurs during execution
    */
   public static QuarkusApp newApplication(
-      AppModel appModel, Path projectRoot, Path targetDirectory, Properties applicationProperties)
+      ApplicationModel appModel,
+      Path projectRoot,
+      Path targetDirectory,
+      Properties applicationProperties)
       throws MojoExecutionException {
     return newApplication(
         appModel,
@@ -145,7 +140,7 @@ public class QuarkusApp implements AutoCloseable {
    * @throws MojoExecutionException if an error occurs during execution
    */
   public static QuarkusApp newApplication(
-      AppModel appModel,
+      ApplicationModel appModel,
       Path projectRoot,
       Path targetDirectory,
       Properties applicationProperties,